Changes in JRuby 9K concerning protected methods
I've removed the expectation on the filter API for the `filter?` and the
`filter_matched` methods, we should not test for internal methods and we
have other test that check for the filter plugin behavior.

in JRuby 9k `respond_to?` behavior has changed, in the following
example:

```
class Hello
  protected
  def name
    "ph"
  end
end

puts Hello.new.respond_to?(:name) == true ? "respond to #name": "doesn't respond_to #name"
```

In JRuby 9k it will return: "Doesn't respond to #name"
In JRuby 1.7.25 it will return "respond_to #name"

This also correspond to the change between MRI 1.9.X and 2.4.x.