Fix and Simplify the offline pack installation
With the upgrade to 9k we also received a new version of the rubygems
library that broke the installation of offline pack. This is due to the
interaction of the bundler loaded gems and the Local index of rubygems
gems.

This commit fix the issue by removing the new to use the local index and
just install the gems in the appropriate directory so bundler can
directly find them.